==> analysis/5-BMDM_PKM2.Rmd <==
---
title: "BMDM and PKM2"
author: "Tess Brodie"
date: University of Bern, DBMR, Visceral and Transplantation Surgery lab
output: 
  html_document:
    toc_depth: 2
---


## BMDM and PKM2
```{r setup, include=FALSE, dev='png'}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(autodep = TRUE, cache = TRUE, cache.lazy = FALSE)
knitr::opts_chunk$set(dev = 'png')
```


### Load packages

```{r load-libs, message = FALSE, warning = FALSE}
library(CATALYST)
library(dplyr)
library(flowCore)
library(ggplot2)
library(diffcyt)
library(cowplot)
library(viridis)
library(RColorBrewer)
library(workflowr)
```

### Load data
```{r}
sce <- readRDS(file.path("output", "sce.rds"))
```


## Differential State analysis
We see that the BMDM have increasing amounts of PKM2 over time after PhX and we want to see if the increase in PKM2 expression is associated with other significant phenotypic changes in this subset. 

# DS analysis
```{r DS_BMDM}
# create design matrix
design <- createDesignMatrix(ei(sce), "condition")
colnames(design) <- gsub("condition", "", colnames(design))

# run DS analysis for control vs. each timepoint
for (i in colnames(design)[-1]) {
  # create contrast matrix
  contrast <- createContrast(as.numeric(colnames(design) == i))
    
  # run DS analysis & extract results table
  res_DS <- diffcyt(sce, clustering_to_use = "mm",
                  analysis_type = "DS", method_DS = "diffcyt-DS-limma",
                  design = design, contrast = contrast, verbose = FALSE)

  # extract results table
  tbl_DS <- rowData(res_DS$res)
  
   # include all results for selected cluster
  k <- metadata(res_DS$res)$clustering_name
  sub1 <- filterSCE(sce, cluster_id == "BMDM", k = k)
  

  # write table to .csv
    fn <- file.path("output", sprintf("DS-results_%s.csv", i))
    write.csv(tbl_DS, fn, row.names = FALSE)
    

   # subset control and timepoint of interest
    sub <- filterSCE(sub1, condition %in% c("control", !!i))
    
    
  # plot differential heatmap
    p <- plotDiffHeatmap(sub, tbl_DS, all = TRUE, col_anno = "condition")
    cat("## ", i, "\n"); print(p); cat("\n\n")
} 
```

### Plot all timepoints and significant changes in BMDM
This needs editing because now it is showing all timepoints, but the significance shown is for control vs 48 hours.  It is very tricky to show the p values for each condition in one graph... I'm not sure how that could be done.  Maybe the heatmap of the target expression over time and the fold change is enough and we could report the P values for each timepoint in a table?  Or we compare the control to all the other conditions globally... This is to be discussed.
```{r DS_BMDM_alltimepoints}
# create design matrix
design <- createDesignMatrix(ei(sce), "condition")
colnames(design) <- gsub("condition", "", colnames(design))

# run DS analysis for control vs. each timepoint
for (i in colnames(design)[-1]) 
  # create contrast matrix
  contrast <- createContrast(as.numeric(colnames(design) == i))
    
  # run DS analysis & extract results table
  res_DS <- diffcyt(sce, clustering_to_use = "mm",
                  analysis_type = "DS", method_DS = "diffcyt-DS-limma",
                  design = design, contrast = contrast, verbose = FALSE)

  # extract results table
  tbl_DS <- rowData(res_DS$res)
  
   # include all results for selected cluster
  k <- metadata(res_DS$res)$clustering_name
  sub2 <- filterSCE(sce, cluster_id == "BMDM", k = k)

  # write table to .csv (doesnt work here)
    fn <- file.path("output", sprintf("DS-results_%s.csv", i))
    write.csv(tbl_DS, fn, row.names = FALSE)
    
    
  # plot differential heatmap
    p <- plotDiffHeatmap(sub2, tbl_DS, all = TRUE, col_anno = "condition")
    cat("## ", i, "\n"); print(p); cat("\n\n")
```

### Filter on BMDM 
```{r Filter_BMDM}
sceB <- filterSCE(sce, k = "mm",
               cluster_id %in% "BMDM")

plot_grid(
  plotDR(sce, dr= "TSNE", color_by = "mm"),
  plotDR(sceB, dr= "TSNE", color_by = "mm"))
```

### Target expression in BMDM for M1 and M2 targets

```{r ExprHeatmap_sig_targets}
plotExprHeatmap(sceB, features = c("PKM2", "CD93", "CD80", "Arginase_1", "CD274", "CD86", "iNOS", "CD1d", "CD40", "CD206"),
    row_anno = FALSE,   # don't annotate samples
    row_clust = FALSE,  # keep samples in original order
    col_clust = TRUE,  # keep markers in original order
    bin_anno = FALSE,   # don't annotate bins
    bars = FALSE,       # don't include sample sizes
    scale = "last")     # aggregate, then scale
```

## TSNE of BMDMs colored by condition
```{r TSNE_condition}
sceB <- runDR(sceB, dr = "TSNE", features = NULL)
plotDR(sceB, dr = "TSNE", color_by = "condition")
```

## Facet by condition
```{r TSNE_condition_facet}
plotDR(sceB, dr = "TSNE", color_by = "condition", facet_by = "condition")
```

## BMDM targets that change significantly in TSNE facet
```{r plotDR-expression, fig.width = 12, fig.height = 5}
p <- plotDR(sceB, dr = "TSNE", color_by = c("PKM2", "CD93", "CD115", "CD49b_DX5", "CD161", "CD80", "I_A_I_E", "CD196", "Arginase_1", "CD64", "CD274", "CD86"))
p$facet$params$ncol <- 4; p
```

## Number of BMDM per sample
```{r BMDM_count}
plotCounts(sceB, 
           group_by = "sample_id", 
           color_by = "condition")
```

## Cluster BMDM into 10 subsets
```{r cluster_10, message = FALSE}
sceB <- cluster(sceB, maxK = 10)
```

## Expression heatmap of 10 BMDM subsets
```{r plotExprHeatmap_10, fig.width = 6, fig.height = 5}
plotExprHeatmap(sceB, features = type_markers(sce), 
    by = "cluster_id", k = "meta10", row_anno = FALSE, bars = TRUE)
```

## Expression heatmap with targets for Nicolas 22_07_20
```{r plotExprHeatmap_10_Ntargets, fig.width = 6, fig.height = 5}
plotExprHeatmap(sceB, features = c("CD192", "CD206", "Ly6C", "Gata6", "CD11c", "CD1d", "Cx43", "CD196", "CD64", "Arginase_1", "CD49b_DX5", "I_A_I_E", "iNOS", "CD161", "PKM2"), 
    by = "cluster_id", k = "meta10", row_anno = FALSE, bars = TRUE)
```

## MultiHeatmap with 10 BMDM subsets
```{r MultiHeatmap_10}
plotMultiHeatmap(sceB, 
                 hm1 = "type", hm2 = "state", 
                 k = "meta10", row_anno = FALSE,
                 col_dend = c(FALSE, FALSE))
```

## Heatmap of PKM2 expression in 10 BMDM clusters over time
```{r ExprHeatmap_PKM2}
plotExprHeatmap(sceB, features = "PKM2", by = "both", k = "meta10", 
    scale = "first", col_clust = FALSE, col_anno = FALSE,
    row_anno = FALSE, 
    bars = TRUE)
```

## 10 BMDM clusters overlaid on TSNE plot
```{r TSNE_meta10}
plotDR(sceB, dr = "TSNE", color_by = "meta10")
```

## TSNE of clusters over time for Nicolas 22_07_20
```{r TSNE_meta10_conditions}
plotDR(sceB, dr = "TSNE", color_by = "meta10", facet_by = "condition")
```

## TSNE over time with PKM2 expression
```{r TSNE_PKM2_conditions}
plotDR(sceB, dr = "TSNE", color_by = "PKM2", facet_by = "condition")
```

## Cluster BMDM into 15 subsets
```{r cluster_15, message = FALSE}
sceB15 <- cluster(sceB, maxK = 15)
```

## Expression heatmap with 15 clusters
```{r ExprHeatmap_15, fig.width = 6, fig.height = 5}
plotExprHeatmap(sceB15, features = type_markers(sce), 
   by = "cluster_id", k = "meta15", row_anno = FALSE, bars = TRUE)
```

## Multiheatmap with 15 clusters
```{r MultiHeatmap_15}
plotMultiHeatmap(sceB15, 
                 hm1 = "type", hm2 = "state", 
                 k = "meta15", row_anno = FALSE,
                 col_dend = c(FALSE, FALSE))
```

## 15 BMDM clusters overlaid on TSNE plot
```{r TSNE_15}
plotDR(sceB15, dr = "TSNE", color_by = "meta15")
```

## Cluster BMDM into 20 subsets
```{r cluster_20, message = FALSE}
sceB20 <- cluster(sceB, maxK = 20)
```

## Expression heatmap of 20 BMDM subsets
```{r ExprHeatmap_20, fig.width = 6, fig.height = 5}
plotExprHeatmap(sceB20, features = type_markers(sce), 
   by = "cluster_id", k = "meta20", row_anno = FALSE, bars = TRUE)
```

## Plot multiheatmap with 20 BMDM subsets
```{r MultiHeatmap_20}
plotMultiHeatmap(sceB20, 
                 hm1 = "type", hm2 = "state", 
                 k = "meta20", row_anno = FALSE,
                 col_dend = c(FALSE, FALSE))
```

## 20 BMDM clusters overlaid on TSNE plot
```{r TSNE_20}
plotDR(sceB20, dr = "TSNE", color_by = "meta20")
```

## Cluster BMDM into 25 subsets
```{r cluster_25, message = FALSE}
sceB25 <- cluster(sceB, maxK = 25)
```

## Expression heatmap of 25 BMDM subsets
```{r ExprHeatmap_25, fig.width = 6, fig.height = 5}
plotExprHeatmap(sceB25, features = type_markers(sce), 
    by = "cluster_id", k = "meta25", row_anno = FALSE, bars = TRUE)
```

## MultiHeatmap with 25 subsets
```{r MultiHeatmap_25}
plotMultiHeatmap(sceB25, 
                 hm1 = "type", hm2 = "state", 
                 k = "meta25", row_anno = FALSE,
                 col_dend = c(FALSE, FALSE))
```

## 25 BMDM clusters overlaid on TSNE plot
```{r TSNE_25}
plotDR(sceB25, dr = "TSNE", color_by = "meta25")
```

## Differential Abundances analysis with all timepoints shown for Nicolas 22_07_20
The significance is shown for control vs 48 hours here
```{r da_all}
# create design matrix
design <- createDesignMatrix(ei(sceB), "condition")
colnames(design) <- gsub("condition", "", colnames(design))

# run DA analysis for control vs. each timepoint
for (i in colnames(design)[-1]) 
  # create contrast matrix
  contrast <- createContrast(as.numeric(colnames(design) == i))
    
  # run DA analysis & extract results table
 res_DA <- diffcyt(sceB, clustering_to_use = "meta10",
                  analysis_type = "DA", method_DA = "diffcyt-DA-edgeR",
                  design = design, contrast = contrast, verbose = FALSE)

  # extract results table
  tbl_DA <- rowData(res_DA$res)

  # write table to .csv (doesnt work here)
    fn <- file.path("output", sprintf("DA-results_%s.csv", i))
    write.csv(tbl_DA, fn, row.names = FALSE)
    
    
  # plot differential heatmap
    plotDiffHeatmap(sceB, tbl_DA,top_n = 20, all = TRUE, col_anno = "condition")
```

## Another option to show cluster abundance
This is for if the significance part is not important and you want the clusters grouped by similarity to eachother
```{r FreqHeatmap_BMDM10}
plotFreqHeatmap(sceB, 
    k = "meta10",
    hm_pal = rev(hcl.colors(10, "RdBu")),
    k_pal = hcl.colors(7, "Zissou 1"),
    col_clust = FALSE, col_anno = FALSE,
    row_anno = FALSE,
    bars = TRUE, perc = TRUE)
```

## Differential state Analysis for all 10 BMDM clusters
```{r DS}
# create design matrix
design <- createDesignMatrix(ei(sceB), "condition")
colnames(design) <- gsub("condition", "", colnames(design))

# run DS analysis for control vs. each timepoint
for (i in colnames(design)[-1]) {
  # create contrast matrix
  contrast <- createContrast(as.numeric(colnames(design) == i))
    
  # run DS analysis & extract results table
  res_DS <- diffcyt(sceB, clustering_to_use = "meta10",
                  analysis_type = "DS", method_DS = "diffcyt-DS-limma",
                  design = design, contrast = contrast, verbose = FALSE)

  # extract results table
  tbl_DS <- rowData(res_DS$res)

  # write table to .csv
    fn <- file.path("output", sprintf("DS-resultsBMDM_%s.csv", i))
    write.csv(tbl_DS, fn, row.names = FALSE)
    
  # subset control and timepoint of interest
    sub <- filterSCE(sceB, condition %in% c("control", !!i))
    
  # plot differential heatmap
    p <- plotDiffHeatmap(sub, tbl_DS, all = TRUE, col_anno = "condition")
    cat("## ", i, "\n"); print(p); cat("\n\n")
} 
```

## Differential state filtered on PKM2
```{r DS_PKM2}
# create design matrix
design <- createDesignMatrix(ei(sceB), "condition")
colnames(design) <- gsub("condition", "", colnames(design))

# run DS analysis for control vs. each timepoint
for (i in colnames(design)[-1]) {
  # create contrast matrix
  contrast <- createContrast(as.numeric(colnames(design) == i))
    
  # run DS analysis & extract results table
  res_DS <- diffcyt(sceB, clustering_to_use = "meta10",
                  analysis_type = "DS", method_DS = "diffcyt-DS-limma",
                  design = design, contrast = contrast, verbose = FALSE)

  # extract results table
  tbl_DS <- rowData(res_DS$res)

  # write table to .csv
    fn <- file.path("output", sprintf("DS-resultsBMDM_%s.csv", i))
    write.csv(tbl_DS, fn, row.names = FALSE)
    
  # subset control and timepoint of interest
    sub <- filterSCE(sceB, condition %in% c("control", !!i))
    
  # plot differential heatmap for PKM2
    p <- plotDiffHeatmap(sub["PKM2", ], tbl_DS, all = TRUE, col_anno = FALSE)
    cat("## ", i, "\n"); print(p); cat("\n\n")
} 
```

## Differential Abundances analysis with 10 BMDM subsets

# DA analysis {.tabset}
```{r da, results = "asis", fig.width = 8, fig.height = 6}
# create design matrix
design <- createDesignMatrix(ei(sceB), "condition")
colnames(design) <- gsub("condition", "", colnames(design))

# run DA analysis for control vs. each timepoint
for (i in colnames(design)[-1]) {
    # create contrast matrix
    contrast <- createContrast(as.numeric(colnames(design) == i))
    
    # run DA analysis
    res <- diffcyt(sceB, clustering_to_use = "meta10",
        analysis_type = "DA", method_DA = "diffcyt-DA-edgeR",
        design = design, contrast = contrast, verbose = FALSE)
    
    # extract results table
    tbl <- rowData(res$res)
    
    # write table to .csv
    fn <- file.path("output", sprintf("DA-resultsBMDM_%s.csv", i))
    write.csv(tbl, fn, row.names = FALSE)
    
    # subset control and timepoint of interest
    sub <- filterSCE(sceB, condition %in% c("control", !!i))
    
    # plot differential heatmap
    p <- plotDiffHeatmap(sub, tbl, all = TRUE, col_anno = "condition")
    cat("## ", i, "\n"); print(p); cat("\n\n")
}
```
